refactor: move http.js from service to network
